requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.8
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.11.0
Faker==11.3.0
gunicorn==20.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg_resources==0.0.0
psycopg2-binary==2.8.6
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
validate-docbr==1.10.0
whitenoise==5.3.0
wrapt==1.12.1
